refactor: move flux service code to another module
